+++ admin-integration-service/src/main/java/com/linki/admin_integration_service/common/redirect/repository/RedirectRepository.java
@@ -11,5 +11,4 @@ public interface RedirectRepository extends JpaRepository<RedirectLinks, String>
     @Query("SELECT COUNT(r) FROM RedirectLinks r WHERE r.contract.proposal.influencer.influencerId = :influencerId")
     long countByInfluencerId(@Param("influencerId") String influencerId);
 
-    RedirectLinks findByShortUrl(String shortUrl);
 }
